requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.12
django-heroku==0.3.1
feedparser==5.2.1
gunicorn==19.9.0
idna==2.8
newsapi-python==0.2.5
psycopg2==2.8.4
pytz==2019.3
requests==2.21.0
urllib3==1.24.3
whitenoise==4.1.4
